Module aws_sdk_customerprofiles::error

Modules

access_denied_exception

See AccessDeniedException

bad_request_exception

See BadRequestException

internal_server_exception

See InternalServerException

resource_not_found_exception

See ResourceNotFoundException

throttling_exception

See ThrottlingException

Structs

AccessDeniedException

You do not have sufficient access to perform this action.

AddProfileKeyError

BadRequestException

The input you provided is invalid.

CreateDomainError

CreateProfileError

DeleteDomainError

DeleteIntegrationError

DeleteProfileError

DeleteProfileKeyError

DeleteProfileObjectError

DeleteProfileObjectTypeError

GetDomainError

GetIntegrationError

GetMatchesError

GetProfileObjectTypeError

GetProfileObjectTypeTemplateError

InternalServerException

An internal service error occurred.

ListAccountIntegrationsError

ListDomainsError

ListIntegrationsError

ListProfileObjectTypeTemplatesError

ListProfileObjectTypesError

ListProfileObjectsError

ListTagsForResourceError

MergeProfilesError

PutIntegrationError

PutProfileObjectError

PutProfileObjectTypeError

ResourceNotFoundException

The requested resource does not exist, or access was denied.

SearchProfilesError

TagResourceError

ThrottlingException

You exceeded the maximum number of requests.

UntagResourceError

UpdateDomainError

UpdateProfileError

Enums

AddProfileKeyErrorKind

CreateDomainErrorKind

CreateProfileErrorKind

DeleteDomainErrorKind

DeleteIntegrationErrorKind

DeleteProfileErrorKind

DeleteProfileKeyErrorKind

DeleteProfileObjectErrorKind

DeleteProfileObjectTypeErrorKind

GetDomainErrorKind

GetIntegrationErrorKind

GetMatchesErrorKind

GetProfileObjectTypeErrorKind

GetProfileObjectTypeTemplateErrorKind

ListAccountIntegrationsErrorKind

ListDomainsErrorKind

ListIntegrationsErrorKind

ListProfileObjectTypeTemplatesErrorKind

ListProfileObjectTypesErrorKind

ListProfileObjectsErrorKind

ListTagsForResourceErrorKind

MergeProfilesErrorKind

PutIntegrationErrorKind

PutProfileObjectErrorKind

PutProfileObjectTypeErrorKind

SearchProfilesErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateDomainErrorKind

UpdateProfileErrorKind