Module aws_sdk_customerprofiles::model

Modules

address

See Address

connector_operator

See ConnectorOperator

domain_stats

See DomainStats

field_source_profile_ids

See FieldSourceProfileIds

flow_definition

See FlowDefinition

incremental_pull_config

See IncrementalPullConfig

list_domain_item

See ListDomainItem

list_integration_item

See ListIntegrationItem

list_profile_object_type_item

See ListProfileObjectTypeItem

list_profile_object_type_template_item

See ListProfileObjectTypeTemplateItem

list_profile_objects_item

See ListProfileObjectsItem

marketo_source_properties

See MarketoSourceProperties

match_item

See MatchItem

matching_request

See MatchingRequest

matching_response

See MatchingResponse

object_filter

See ObjectFilter

object_type_field

See ObjectTypeField

object_type_key

See ObjectTypeKey

profile

See Profile

s3_source_properties

See S3SourceProperties

salesforce_source_properties

See SalesforceSourceProperties

scheduled_trigger_properties

See ScheduledTriggerProperties

service_now_source_properties

See ServiceNowSourceProperties

source_connector_properties

See SourceConnectorProperties

source_flow_config

See SourceFlowConfig

task

See Task

trigger_config

See TriggerConfig

trigger_properties

See TriggerProperties

update_address

See UpdateAddress

zendesk_source_properties

See ZendeskSourceProperties

Structs

Address

A generic address associated with the customer that is not mailing, shipping, or

ConnectorOperator

The operation to be performed on the provided source fields.

DomainStats

Usage-specific statistics about the domain.

FieldSourceProfileIds

A duplicate customer profile that is to be merged into a main profile.

FlowDefinition

The configurations that control how Customer Profiles retrieves data from the source,

IncrementalPullConfig

Specifies the configuration used when importing incremental records from the

ListDomainItem

An object in a list that represents a domain.

ListIntegrationItem

An integration in list of integrations.

ListProfileObjectTypeItem

A ProfileObjectType instance.

ListProfileObjectTypeTemplateItem

A ProfileObjectTypeTemplate in a list of ProfileObjectTypeTemplates.

ListProfileObjectsItem

A ProfileObject in a list of ProfileObjects.

MarketoSourceProperties

The properties that are applied when Marketo is being used as a source.

MatchItem

The Match group object.

MatchingRequest

The flag that enables the matching process of duplicate profiles.

MatchingResponse

The flag that enables the matching process of duplicate profiles.

ObjectFilter

The filter applied to ListProfileObjects response to include profile objects with the specified index values.

ObjectTypeField

Represents a field in a ProfileObjectType.

ObjectTypeKey

An object that defines the Key element of a ProfileObject. A Key is a special element

Profile

The standard profile of a customer.

S3SourceProperties

The properties that are applied when Amazon S3 is being used as the flow source.

SalesforceSourceProperties

The properties that are applied when Salesforce is being used as a source.

ScheduledTriggerProperties

Specifies the configuration details of a scheduled-trigger flow that you define.

ServiceNowSourceProperties

The properties that are applied when ServiceNow is being used as a source.

SourceConnectorProperties

Specifies the information that is required to query a particular Amazon AppFlow connector.

SourceFlowConfig

Contains information about the configuration of the source connector used in the

Task

A class for modeling different type of tasks. Task implementation varies based on the

TriggerConfig

The trigger settings that determine how and when Amazon AppFlow runs the specified

TriggerProperties

Specifies the configuration details that control the trigger for a flow. Currently,

UpdateAddress

Updates associated with the address properties of a customer profile.

ZendeskSourceProperties

The properties that are applied when using Zendesk as a flow source.

Enums

DataPullMode

FieldContentType

Gender

MarketoConnectorOperator

OperatorPropertiesKeys

PartyType

S3ConnectorOperator

SalesforceConnectorOperator

ServiceNowConnectorOperator

SourceConnectorType

StandardIdentifier

TaskType

TriggerType

ZendeskConnectorOperator