Module aws_sdk_customerprofiles::operation

Structs

AddProfileKey

Associates a new key value with a specific profile, such as a Contact Trace Record (CTR)

CreateDomain

Creates a domain, which is a container for all customer data, such as customer profile

CreateProfile

Creates a standard profile.

DeleteDomain

Deletes a specific domain and all of its customer data, such as customer profile

DeleteIntegration

Removes an integration from a specific domain.

DeleteProfile

Deletes the standard customer profile and all data pertaining to the profile.

DeleteProfileKey

Removes a searchable key from a customer profile.

DeleteProfileObject

Removes an object associated with a profile of a given ProfileObjectType.

DeleteProfileObjectType

Removes a ProfileObjectType from a specific domain as well as removes all the

GetDomain

Returns information about a specific domain.

GetIntegration

Returns an integration for a domain.

GetMatches

This API is in preview release for Amazon Connect and subject to change.

GetProfileObjectType

Returns the object types for a specific domain.

GetProfileObjectTypeTemplate

Returns the template information for a specific object type.

ListAccountIntegrations

Lists all of the integrations associated to a specific URI in the AWS account.

ListDomains

Returns a list of all the domains for an AWS account that have been created.

ListIntegrations

Lists all of the integrations in your domain.

ListProfileObjectTypeTemplates

Lists all of the template information for object types.

ListProfileObjectTypes

Lists all of the templates available within the service.

ListProfileObjects

Returns a list of objects associated with a profile of a given ProfileObjectType.

ListTagsForResource

Displays the tags associated with an Amazon Connect Customer Profiles resource. In Connect

MergeProfiles

This API is in preview release for Amazon Connect and subject to change.

PutIntegration

Adds an integration between the service and a third-party service, which includes

PutProfileObject

Adds additional objects to customer profiles of a given ObjectType.

PutProfileObjectType

Defines a ProfileObjectType.

SearchProfiles

Searches for profiles within a specific domain name using name, phone number, email

TagResource

Assigns one or more tags (key-value pairs) to the specified Amazon Connect Customer Profiles

UntagResource

Removes one or more tags from the specified Amazon Connect Customer Profiles resource. In Connect

UpdateDomain

Updates the properties of a domain, including creating or selecting a dead letter queue

UpdateProfile

Updates the properties of a profile. The ProfileId is required for updating a customer